chore(auth): add paper-desk variant controls for all auth pages
Rename the auth-flow variant redesign-2026-06-02 -> paper-desk and bring
email verification into the variant system, so login, signup, invited
signup and email verification all resolve legacy vs paper-desk through
authFlowVariantRegistry. Paper-desk pages remain stubs; real designs land
in follow-up PRs.
