+++ frontend/src/scenes/authentication/authFlowVariants.ts
@@ -1,11 +1,11 @@
 import { FEATURE_FLAGS } from 'lib/constants'
 import type { FeatureFlagsSet } from 'lib/logic/featureFlagLogic'
 
-export type AuthFlowVariant = 'legacy' | 'redesign-2026-06-02'
+export type AuthFlowVariant = 'legacy' | 'paper-desk'
 
 const DEFAULT_VARIANT: AuthFlowVariant = 'legacy'
 
-export const AUTH_FLOW_VARIANTS: AuthFlowVariant[] = ['legacy', 'redesign-2026-06-02']
+export const AUTH_FLOW_VARIANTS: AuthFlowVariant[] = ['legacy', 'paper-desk']
 
 export function resolveAuthFlowVariant(featureFlags: FeatureFlagsSet): AuthFlowVariant {
     const variant = featureFlags[FEATURE_FLAGS.AUTH_FLOW_VARIANT]
